Crate penumbra_sdk_transaction

Data modeling for Penumbra transactions.

This crate defines data structures that provide modeling of shielded transactions through their entire lifecycle:

the TransactionPlan type completely describes a planned transaction before it is created;

the Transaction type represents the shielded transaction itself;

the TransactionView type represents a view from a particular TransactionPerspective (e.g., the sender or receiver) of the cleartext contents of a shielded transaction after it has been created.

Re-exports

pub use action::Action;

pub use action_list::ActionList;

pub use plan::ActionPlan;

pub use plan::TransactionPlan;

pub use view::ActionView;

pub use view::MemoPlaintextView;

pub use view::MemoView;

pub use view::TransactionPerspective;

pub use view::TransactionView;

pub use penumbra_sdk_txhash as txhash;

Modules

action

action_list

gas

memo

plan

Declarative transaction plans, used for transaction authorization and creation.

view

Structs

AuthorizationData

Authorization data returned in response to a TransactionDescription.

Compat

A compatibility wrapper for trait implementations that are temporarily duplicated in multiple crates as an orphan rule work around until we finish splitting crates (#2288).

DetectionData

Detection data used by a detection server using Fuzzy Message Detection.

Transaction

TransactionBody

TransactionParameters

Parameters determining when the transaction should be accepted to the chain.

WitnessData

Enums

Error

Traits

IsAction

Common behavior between Penumbra actions.