penumbra_sdk_transaction
Module plan

Declarative transaction plans, used for transaction authorization and creation.

Structs

CluePlan

DetectionDataPlan

MemoPlan

TransactionPlan

A declaration of a planned Transaction, for use in transaction authorization and creation.

Enums

ActionPlan

A declaration of a planned Action, for use in transaction creation.