penumbra_sdk_transaction::gas
Trait GasCost

pub trait GasCost {
    // Required method
    fn gas_cost(&self) -> Gas;
}

Allows Actions and Transactions to statically indicate their relative resource consumption. Since the gas cost needs to be multiplied by a price, the values returned only need to be scaled relatively to each other.

Required Methods

fn gas_cost(&self) -> Gas